+++ lua/obsidian/builtin.lua
@@ -7,7 +7,7 @@ local util = require "obsidian.util"
 M.smart_action = function()
   local legacy = Obsidian.opts.legacy_commands
   -- follow link if possible
-  if api.cursor_on_markdown_link(nil, nil, true) then
+  if api.cursor_link() then
     return legacy and "<cmd>ObsidianFollowLink<cr>" or "<cmd>Obsidian follow_link<cr>"
   end