+++ lua/obsidian/commands/init-legacy.lua
@@ -1,4 +1,5 @@
 local util = require "obsidian.util"
+local search = require "obsidian.search"
 local iter = vim.iter
 
 local command_lookups = {
@@ -110,7 +111,7 @@ M.complete_args_search = function(client, _, cmd_line, _)
 
   local completions = {}
   local query_lower = string.lower(query)
-  for note in iter(client:find_notes(query, { search = { sort = true } })) do
+  for note in iter(search.find_notes(query, { search = { sort = true } })) do
     local note_path = assert(note.path:vault_relative_path { strict = true })
     if string.find(string.lower(note:display_name()), query_lower, 1, true) then
       table.insert(completions, note:display_name() .. "  " .. tostring(note_path))