+++ lua/obsidian/commands/open.lua
@@ -1,6 +1,7 @@
 local api = require "obsidian.api"
-local RefTypes = require("obsidian.search").RefTypes
 local Path = require "obsidian.path"
+local search = require "obsidian.search"
+local log = require "obsidian.log"
 
 ---@param path? string|obsidian.Path
 local function open_in_app(path)
@@ -26,13 +27,13 @@ local function open_in_app(path)
   else
     uri = ("obsidian://open?vault=%s&file=%s"):format(encoded_vault, encoded_path)
   end
+  print(uri)
 
   Obsidian.opts.open.func(uri)
 end
 
----@param client obsidian.Client
 ---@param data CommandArgs
-return function(client, data)
+return function(_, data)
   ---@type string|?
   local search_term
 
@@ -40,19 +41,19 @@ return function(client, data)
     search_term = data.args
   else
     -- Check for a note reference under the cursor.
-    local cursor_link, _, ref_type = api.parse_cursor_link()
-    if cursor_link ~= nil and ref_type ~= RefTypes.NakedUrl and ref_type ~= RefTypes.FileUrl then
-      search_term = cursor_link
-    end
+    local link_string, _ = api.cursor_link()
+    search_term = link_string
   end
 
   if search_term then
-    -- Try to resolve search term to a single note.
-    client:resolve_note_async_with_picker_fallback(search_term, function(note)
+    search.resolve_link_async(search_term, function(results)
+      if vim.tbl_isempty(results) then
+        return log.err "Note under cusros is not resolved"
+      end
       vim.schedule(function()
-        open_in_app(note.path)
+        open_in_app(results[1].path)
       end)
-    end, { prompt_title = "Select note to open" })
+    end)
   else
     -- Otherwise use the path of the current buffer.
     local bufname = vim.api.nvim_buf_get_name(0)