+++ tests/test_api.lua
@@ -59,47 +59,26 @@ T["toggle_checkbox"]["should use custom states if provided"] = function()
   eq("- [!] dummy", child.api.nvim_get_current_line())
 end
 
-T["cursor_on_markdown_link"] = function()
+T["cursor_link"] = function()
   --                                               0    5    10   15   20   25   30   35   40    45  50   55
   --                                               |    |    |    |    |    |    |    |    |    |    |    |
   child.api.nvim_buf_set_lines(0, 0, -1, false, { "The [other](link/file.md) plus [[another/file.md|yet]] there" })
 
+  local link1 = "[other](link/file.md)"
+  local link2 = "[[another/file.md|yet]]"
+
   local tests = {
-    { cur_col = 3, open = nil, close = nil },
-    { cur_col = 4, open = 5, close = 25 },
-    { cur_col = 6, open = 5, close = 25 },
-    { cur_col = 24, open = 5, close = 25 },
-    { cur_col = 25, open = nil, close = nil },
-    { cur_col = 30, open = nil, close = nil },
-    { cur_col = 31, open = 32, close = 54 },
-    { cur_col = 39, open = 32, close = 54 },
-    { cur_col = 53, open = 32, close = 54 },
-    { cur_col = 54, open = nil, close = nil },
+    { cur_col = 4, link = link1, t = "Markdown" },
+    { cur_col = 6, link = link1, t = "Markdown" },
+    { cur_col = 24, link = link1, t = "Markdown" },
+    { cur_col = 31, link = link2, t = "WikiWithAlias" },
+    { cur_col = 39, link = link2, t = "WikiWithAlias" },
+    { cur_col = 53, link = link2, t = "WikiWithAlias" },
   }
   for _, test in ipairs(tests) do
     child.api.nvim_win_set_cursor(0, { 1, test.cur_col })
-    local open, close = unpack(child.lua [[local open, close = M.cursor_on_markdown_link(); return { open, close }]])
-    eq(test.open, open)
-    eq(test.close, close)
-  end
-end
-
-T["parse_cursor_link"] = function()
-  child.api.nvim_buf_set_lines(0, 0, -1, false, { "The [other](link/file.md) plus [[another/file.md|yet]] there" })
-
-  local tests = {
-    { cur_col = 4, loc = "link/file.md", name = "other", t = "Markdown" },
-    { cur_col = 6, loc = "link/file.md", name = "other", t = "Markdown" },
-    { cur_col = 24, loc = "link/file.md", name = "other", t = "Markdown" },
-    { cur_col = 31, loc = "another/file.md", name = "yet", t = "WikiWithAlias" },
-    { cur_col = 39, loc = "another/file.md", name = "yet", t = "WikiWithAlias" },
-    { cur_col = 53, loc = "another/file.md", name = "yet", t = "WikiWithAlias" },
-  }
-  for _, test in ipairs(tests) do
-    child.api.nvim_win_set_cursor(0, { 1, test.cur_col })
-    local loc, name, t = unpack(child.lua [[local loc, name, t = M.parse_cursor_link(); return { loc, name, t }]])
-    eq(test.loc, loc)
-    eq(test.name, name)
+    local link, t = unpack(child.lua_get [[{ M.cursor_link() }]])
+    eq(test.link, link)
     eq(test.t, t)
   end
 end