+++ CHANGELOG.md
@@ -51,6 +51,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved `client:apply_async` -> `api.dir`
 - Deprecate `statusline` options and `vim.g.obsidian` in favor of `footer` options and `vim.b.obsidian_footer`.
 - Moved `client:switch_workspace` -> `Workspace.switch`
+- Moved `client:follow_link_async` -> `api.follow_link`
+- Moved `client:resolve_note_async` -> `search.resolve_note_async`
+- Moved `client:find_notes_async` -> `search.find_notes_async`
 
 ### Fixed